Add check in play game logic and question model change

+++ Bakis/Dtos/Sights/UpdateSightDto.cs
@@ -10,7 +10,7 @@ namespace Bakis.Dtos.Sights
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Points { get; set; }
+        public int Points { get; set; }
         public string Address { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
